seperated addToFavorites and removeFromFavorites functions into it's own file called AddFavorite.jsx for ease of use with other pages that may need it
