+++ src/layout/Home.jsx
@@ -1,28 +1,19 @@
 import { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
 import { useAuth } from "../auth/AuthContext";
 import { useApi } from "../api/ApiContext";
 import useQuery from "../api/useQuery";
-import useMutation from "../api/useMutation";
 import RecipeCard from "./RecipeCard";
 
 const HomePage = () => {
-  const [favoritingRecipe, setFavoritingRecipe] = useState(null);
   const [userFavorites, setUserFavorites] = useState([]);
-  const navigate = useNavigate();
   const { token } = useAuth();
-  const { request, invalidateTags } = useApi();
+  const { request } = useApi();
 
   const {
     data: recipes = [],
     loading,
     error,
   } = useQuery("/favorites/top-favorited", "topRecipes");
-
-  const { mutate: addToFavorites } = useMutation("POST", "/favorites", [
-    "topRecipes",
-    "userFavorites",
-  ]);
 
   useEffect(() => {
     const fetchUserFavorites = async () => {
@@ -44,74 +35,17 @@ const HomePage = () => {
     fetchUserFavorites();
   }, [token, request]);
 
-  const getUserIdFromToken = () => {
-    if (!token) return null;
-
-    try {
-      const payload = JSON.parse(atob(token.split(".")[1]));
-      return payload.user_id || payload.id;
-    } catch (error) {
-      console.error("Error decoding token:", error);
-      return null;
+  const handleFavoriteChange = (recipeId, isNowFavorited) => {
+    if (isNowFavorited) {
+      setUserFavorites((prev) => [...prev, recipeId]);
+    } else {
+      setUserFavorites((prev) => prev.filter((id) => id !== recipeId));
     }
   };
 
-  const handleAddToFavorites = async (recipeId) => {
-    try {
-      setFavoritingRecipe(recipeId);
-
-      const userId = getUserIdFromToken();
-      if (!userId) {
-        navigate("/login");
-        return;
-      }
-
-      const success = await addToFavorites({
-        recipe_id: recipeId,
-        user_id: userId,
-      });
-
-      if (success) {
-        setUserFavorites((prev) => [...prev, recipeId]);
-        alert("Recipe added to favorites!");
-      }
-    } catch (err) {
-      const errorMessage = err.message || err.error || "An error occurred";
-      alert(`Error: ${errorMessage}`);
-      console.error("Full error:", err);
-    } finally {
-      setFavoritingRecipe(null);
-    }
-  };
-
-  const handleRemoveFromFavorites = async (recipeId) => {
-    try {
-      setFavoritingRecipe(recipeId);
-
-      const userId = getUserIdFromToken();
-      if (!userId) {
-        navigate("/login");
-        return;
-      }
-      
-      const result = await request(`/favorites/${recipeId}`, {
-        method: "DELETE",
-        body: JSON.stringify({ user_id: userId }),
-      });
-
-      if (result) {
-        setUserFavorites((prev) => prev.filter((id) => id !== recipeId));
-        invalidateTags(["topRecipes", "userFavorites"]);
-        alert("Recipe removed from favorites!");
-      }
-    } catch (err) {
-      const errorMessage = err.message || err.error || "An error occurred";
-      alert(`Error: ${errorMessage}`);
-      console.error("Full error:", err);
-    } finally {
-      setFavoritingRecipe(null);
-    }
-  };
+  // Calculate max favorites for top favorite display
+  const maxFavorites =
+    recipes.length > 0 ? Math.max(...recipes.map((r) => r.favoritecount)) : 0;
 
   return (
     <div>
@@ -126,21 +60,24 @@ const HomePage = () => {
 
           {recipes.length > 0 && !loading && (
             <div className="recipes-grid">
-              {recipes.map((recipe, index) => (
-                <RecipeCard
-                  key={recipe.id || index}
-                  recipe={recipe}
-                  rank={index + 1}
-                  recipes={recipes}
-                  onAddToFavorites={handleAddToFavorites}
-                  onRemoveFromFavorites={handleRemoveFromFavorites}
-                  favoritingRecipe={favoritingRecipe}
-                  token={token}
-                  showRank={true}
-                  showTopFavoriteTag={true}
-                  isFavorited={userFavorites.includes(recipe.id)}
-                />
-              ))}
+              {recipes.map((recipe, index) => {
+                const isTopFavorited = recipe.favoritecount === maxFavorites;
+
+                return (
+                  <div key={recipe.id || index} className="recipe-card-wrapper">
+                    <div className="recipe-rank">#{index + 1}</div>
+                    {isTopFavorited && (
+                      <div className="top-favorite-tag">🏆 Most Favorited!</div>
+                    )}
+
+                    <RecipeCard
+                      recipe={recipe}
+                      onFavoriteChange={handleFavoriteChange}
+                      isFavorited={userFavorites.includes(recipe.id)}
+                    />
+                  </div>
+                );
+              })}
             </div>
           )}
