+++ src/layout/RecipeCard.jsx
@@ -1,17 +1,7 @@
 import { useNavigate } from "react-router-dom";
+import AddFavorite from "./AddFavorite";
 
-const RecipeCard = ({
-  recipe,
-  rank,
-  recipes,
-  onAddToFavorites,
-  onRemoveFromFavorites,
-  favoritingRecipe,
-  token,
-  showRank = false,
-  showTopFavoriteTag = false,
-  isFavorited = false,
-}) => {
+const RecipeCard = ({ recipe, onFavoriteChange, isFavorited = false }) => {
   const navigate = useNavigate();
 
   const handleUsernameClick = (username) => {
@@ -22,22 +12,12 @@ const RecipeCard = ({
     navigate(`/recipe/${recipe.id}`);
   };
 
-  const maxFavorites = recipes
-    ? Math.max(...recipes.map((r) => r.favoritecount))
-    : 0;
-  const isTopFavorited =
-    showTopFavoriteTag && recipe.favoritecount === maxFavorites;
-
   return (
     <div
       className="recipe-card"
       onClick={handleRecipeClick}
       style={{ cursor: "pointer" }}
     >
-      {showRank && <span>#{rank}</span>}
-      {isTopFavorited && (
-        <span className="top-favorite-tag">🏆 Most Favorited!</span>
-      )}
       <h3>{recipe.name}</h3>
       <p>
         Created by:
@@ -53,28 +33,12 @@ const RecipeCard = ({
       </p>
       <p>Favorites: {recipe.favoritecount}</p>
 
-      {(onAddToFavorites || onRemoveFromFavorites) && (
-        <button
-          onClick={(e) => {
-            e.stopPropagation();
-            if (isFavorited && onRemoveFromFavorites) {
-              onRemoveFromFavorites(recipe.id);
-            } else if (onAddToFavorites) {
-              onAddToFavorites(recipe.id);
-            }
-          }}
-          disabled={favoritingRecipe === recipe.id}
-        >
-          {favoritingRecipe === recipe.id
-            ? isFavorited
-              ? "Removing..."
-              : "Adding..."
-            : !token
-            ? "Login to Favorite"
-            : isFavorited
-            ? "💔 Remove from Favorites"
-            : "❤️ Add to Favorites"}
-        </button>
+      {onFavoriteChange && (
+        <AddFavorite
+          recipeId={recipe.id}
+          isFavorited={isFavorited}
+          onFavoriteChange={onFavoriteChange}
+        />
       )}
     </div>
   );
